notification_service/requirements.txt
flask==3.0.2
twilio==8.12.0
python-dotenv==1.0.1
flask-cors==4.0.0
requests==2.31.0 